### 2024-04-16 #173 @Batalex
---
---

`paper.md` not yet merged, but this is not blocking https://github.com/ekiefl/pooltool/pull/122/files
