### 2024-04-19 #173 @ekiefl
Hi @Batalex, I've made the following changes:

1. I've merged `paper.md`
2. I made a new release (v0.3.2)
3. I made a DOI for v0.3.2: https://doi.org/10.5281/zenodo.10995500

I edited the first post with the bumped version (v0.3.2) and the DOI.

Regarding the survey: I already filled this out, but if your records show otherwise, I'm happy to do it again.

With this all finished, I _think_ we're done with pre-review checks. Please let me know if there's anything else left to do.

Thanks for your time.
