### 2024-05-10 #173 @ekiefl
In anticipating pooltool being reviewed in pyOpenSci, I've given a lot of thought about my licensing and I've decided to switch licenses from GPLv3 to Apache v2. I figured it would be better to do this before any reviews start, so I've made the following changes:

1. Merged the license change (https://github.com/ekiefl/pooltool/pull/123)
2. Made a new release (v0.3.3)
3. I made a DOI for v0.3.3: https://doi.org/10.5281/zenodo.11171124

I edited the first post with the bumped version (v0.3.3) and the DOI.
