### 2024-05-20 #173 @cmarmo
Hi @ekiefl , nice to meet you, and thanks for submitting to pyOpenSci!
I'm Chiara and I'm going to take care of the process as editor.
I'm looking for reviewers right now and hope to be back to you by the end of the week.
