### 2024-06-24 #173 @cmarmo
Hi @ekiefl, I am happy to share that I've found our first reviewer for Pooltool.
@sebastianotronto kindly accepted to review pooltool for pyOpenSci: thank you very much Sebastiano! :pray:
Feel free to introduce yourself here , while I'm pursuing my quest for a second reviewer... :slightly_smiling_face:
