### 2024-07-08 #173 @cmarmo
Hello everybody, I'm waiting for some answers from potential second reviewers.
In general we try to have both reviewers working at the same time, but in order to move forward, and if @sebastianotronto is ready, perhaps we can start with the first review: @ekiefi would this be ok with you?

Thank you very much for your understanding :pray:
