### 2024-07-09 #173 @sebastianotronto
Sure, here is my review! Apologies if I misunderstood some requirements from the standard template.

---

## Review Comments

Here are my general comments. **In bold** the parts that require an answer or action.

I had minor hiccups during the installation because the tool relies on panda3d version 11 (not released yet). The installation instructions are clear, but I tried using an unsupported python version first (3.12) and that did not work out. **It would be nice if the tool could be installed without workaround for panda3d, and possibly on the latest python version. Is it just a matter of waiting for panda3d v11 to be released?**

### Graphical mode

The tool also offers a graphical mode that can be "played" like a 3D game. The overall effect is very realistic and it appears clear that the physics is accurately simulated. I also liked the mouse + keyboard controls, personally I found them intuitive.

The author does not make specific claims about performance, but the graphical mode was quit demanding on my laptop, with extreme lag. On the other hand, it worked like a charm on my much more powerful desktop.

On my laptop I also encountered some graphical glitches (see pictures below):

* Help menu (but also other text) is cut out when playing with unusual resolution (I was keeping two windows side by side)
* Stick disappeared after entering visual mode, only reappears in shot mode or English mode.

![help-cutout](https://github.com/pyOpenSci/software-submission/assets/14557809/855c314c-0d53-43b7-84a9-99874dbcd002)
![no-stick](https://github.com/pyOpenSci/software-submission/assets/14557809/884ca3eb-81fa-4eb6-9b6d-59c110cbc27a)

A couple of times I started a shot that took a very long time to calculate, i.e. more than 10 seconds on my desktop. Both times this happened at the very first shot of the game, default mode. I suspect this is because the starting position of many balls all close together generates a huge number of events compared to other shots (the author explained in his blog that the simulation algorithm is event-based). **I am curious if the author is aware of this and if there are plans to improve the sometimes slow first shot.**

### Documentation and API

The documentation is clear and concise and it makes it easy to get started with the package. It is very much still in the making and some parts are marked as incomplete, but the essentials are there and well written.

The documentation contain an "Hello World" example script to use the APIs, and an in-depth explanation of this example. Very nice way to get started.

The API documentation is automatically generated and it contains plenty of examples. As I noted above, not all functions have examples, but considering how fine-grained the API is I do not think it would be necessary.

I also enjoyed reading the author's blog posts where he explains the physics of the problem and the algorithm used for the simulator. Although not part of the official documentation of the package, it is definitely a nice addition.

The code is generally high-quality and easy to understand, despite the size of the package.

### Possible performance improvement

I have also been thinking about some possible **performance improvement** in the main simulation loop. I don't know if my suggestion would actually lead to any significant improvement, but it is interesting to think about it.

Let's consider the simplified case where there are `n` balls and the only possible events are ball-ball collisions. Right now the main simulation loop would compute, at every iteration, `O(n^2)` possible collisions. I am wondering if it is possible to not re-do every collision check at each iteration, saving some work.

Say for example that at the first iteration it was found that the next collision is between ball 1 and ball 2. At the second iteration, instead of computing all `O(n^2)` ball-ball collisions, we could compute just the `O(n)` collisions involving ball 1 or ball 2. Indeed any two other balls, say for example ball 3 and ball 4, are not affected (yet) by the collision of ball 1 and ball 2, so their next collision time will remain the same computed in the first iteration.

This could cut down significantly the number of computations done at each iteration.
