### 2024-07-13 #173 @ekiefl
@sebastianotronto, I wanted to respond sooner rather than later to let you know that I've read your review, and I'm super appreciative of how thorough it is.

These are great suggestions and I'm going to start making my way through your comments to try and improve some of the areas you've pointed out.
