+++ environment.yml
@@ -1,0 +1,9 @@
+name: pooltool-dev
+channels:
+  - defaults
+  - conda-forge
+dependencies:
+  - python==3.8.10
+  - pip
+  - pip:
+    - poetry
