### 2024-07-14 0cb8b34
Use known python version for architecture
