+++ .github/workflows/lint-and-test.yml
@@ -13,7 +13,7 @@ jobs:
       - name: Install Python
         uses: actions/setup-python@v4
         with:
-          python-version: "3.8.10"
+          python-version: "3.8"
 
       - name: Cache Poetry install
         uses: actions/cache@v2
