### 2024-07-14 0f71b44
Fix poetry run commands in CI
