+++ requirements.txt
@@ -1,15 +1,0 @@
-numpy
-pandas
-numba
-scipy
-panda3d==1.10.13
-panda3d-gltf
-panda3d-simplepbr
-Pillow
-cattrs
-attrs
-pprofile
-msgpack
-msgpack-numpy
-h5py
-pyyaml
