+++ docs/requirements.txt
@@ -1,5 +1,0 @@
-sphinx
-sphinx-copybutton
-furo
-myst-parser
-sphinx-autoapi
