### 2024-07-15 1f209ce
Remove MANIFEST.in and setup.py
