### 2024-07-15 2cd18e8
Fix pooltool API index.rst
