+++ docs/_templates/autoapi/index.rst
@@ -7,8 +7,6 @@ This page contains auto-generated API reference documentation.
    :titlesonly:
    :maxdepth: 1
 
-   {% for page in pages %}
-   {% if page.top_level_object and page.display %}
+   {% for page in pages|selectattr("is_top_level_object") %}
    {{ page.include_path }}
-   {% endif %}
    {% endfor %}
