+++ pyproject.toml
@@ -93,11 +93,11 @@ ruff = "*"
 optional = true
 
 [tool.poetry.group.docs.dependencies]
-sphinx = "*"
-sphinx-copybutton = "*"
-furo = "*"
-myst-parser = "*"
-sphinx-autoapi = "*"
+sphinx = ">=6.1.0"
+sphinx-copybutton = ">=0.5.2"
+furo = ">=2024.5.6"
+myst-parser = ">=3.0.1"
+sphinx-autoapi = ">=3.1.2"
 
 [build-system]
 requires = ["poetry-core"]
