### 2024-07-15 #173 @cmarmo
Dear @ekiefl, I am really happy to announce that we have our second reviewer. :partying_face:
@eliotwrobson kindly accepted to help us as a reviewer: thank you so much! :pray:

As summer is striking in the Northern hemisphere (which is essentially the one involved here), let's say that it would be wonderful to have the review submitted before August 5th. @eliotwrobson, is that ok with you? Thanks for your help!
