### 2024-07-16 #173 @eliotwrobson
@cmarmo I think August 5 sounds reasonable! I'll let you know if anything comes up. My schedule right now is a little in flux but I should be able to make that.
