### 2024-07-17 #173 @eliotwrobson
@ekiefl I've just started poking around the package, looks cool so far! I was having trouble getting the GUI to work on my machine, do you happen to have code that produces the slowdown that @/sebastianotronto pointed out? I'd like to do some profiling.
