### 2024-07-17 #173 @ekiefl
@cmarmo thank you so much for your exhaustive search!

And @eliotwrobson, I really appreciate you agreeing to review. Thank you!

---

> I was having trouble getting the GUI to work on my machine

That's a bit troubling. What are your specs?

> do you happen to have code that produces the slowdown that @/sebastianotronto pointed out?

@sebastianotronto I will respond to your review more fully soon, but since @eliotwrobson is bringing up the performance hits that often occur (several seconds or more), I figured it makes sense to provide some further information for both of you.

While it's definitely possible and probably practical to speed up the algorithm via your suggestion of caching event times to avoid recalculation, I believe the large slowdowns you sometimes observe have to with the package numba.

Numba offers just-in-time compiling of python code into LLVM machine code, which offers performance speeds similar to what one would expect with a low level language like C or Rust or something. And pooltool uses numba extensively in its shot evolution algorithm (probably in around 20+ functions). This is great, but since these functions are just-in-time compiled, the first call is always slow because the function is compiled on the spot. Subsequent calls during that python process are fast because the function is compiled.

Fortunately, numba offers the ability to cache the compiled functions, so that a function compiled on a Tuesday can be re-used on a Wednesday. Pooltool makes full use of this feature to try and ameliorate the problem of re-compiling every time someone boots up a pooltool script. This saves a ton of time, however there is still a significant overhead just to load the cached code.

**TLDR**: numba can cause the first shot to be slow either because the numba functions needs to compiled (really slow) or they've previously been compiled but they still need to be loaded from cache (just kind of slow). Pile on the fact that the break shot is intrinsically slower because there are more events to calculate, and we can end up with some long wait times.

**Partial solution**: While numba compilation and cache loading is inevitable, I think the UX could be improved by calculating an "offscreen" shot as soon as the GUI is opened and introducing a "please wait" screen while the numba's housekeeping business is occurring.

---

@eliotwrobson, now that you have more context, if you are still interested in profiling the code, you might try poking around in `sandbox/break.py` for inspiration. For example, this will calculate 20 breaks and calculate the average time taken for the shot evolution algorithm:

```bash
python sandbox/break.py --no-viz --time-it
```

If you like profiling with cachegrind files, you can create a cachegrind of a break shot with

```python
python sandbox/break.py --no-viz --profile-it
```

In the script you can see lines where I "burn" a break shot to avoid numba compilation/cache loading from being included in the timing/profiling:

```python
        # Burn a run (numba cache loading)
        pt.simulate(shot, continuous=True)
```
