### 2024-07-18 #173 @eliotwrobson
@ekiefl Thank you for your fast response! I ran the command you sent and I think you're right. Even on my system, the output of the benchmarks came out to fractions of a second, so most of the time probably is just cache retrieval. The flamegraph I got is below:

![flamegraph](https://github.com/user-attachments/assets/b2897064-edd1-4f48-8681-459d3b864f52)

I was able to get this during a cached run, and it seems to confirm the discussion on the blog, that most of the time is spent computing all of the events that can happen and finding the minimum. If performance becomes an issue, I think this is a very straightforward option to parallelize using numba, but this might not become relevant unless you want to add better support for larger tables with more balls or something. Overall, I think that having a loading step before opening the GUI is probably the way to go, and anyone running large simulations from a program will just have the first run take slightly longer.

Also, I think I configured something weird on my terminal, so I wouldn't worry too much about it not working for me. Since everything seems alright, I'm going to start going down my checklist 👍🏽
