### 2024-07-21 945a399
run_pooltool -> main.py
When packaged, poetry will wrap up main.py into a CLI tool
