### 2024-07-21 9e50379
Call main.py's run as the CLI `run-pooltool`
