### 2024-07-22 8d0b1b1
Improve the include/exclude pyproject fields
