+++ environment.yml
@@ -6,4 +6,4 @@ dependencies:
   - python==3.8.10
   - pip
   - pip:
-    - poetry
+    - poetry>=1.8.3
