### 2024-07-22 bc832a2
Use dynamic versioning
