### 2024-07-22 eb17e9c
Poetry expects the PyPI name for publishing
