+++ PYPI.md
@@ -4,23 +4,48 @@ These are instructions for how to make new pip-installable pooltool versions on 
 
 ## 1. House keeping
 
-- Make sure `pooltool.__init__.py` has version X.X.X.
-
-- If you added data files since the last release, make sure they have been added to the `pyproject.toml.
-
 - Update `logo.png` if its not a sub-version (rename logo_small.png to previous version (e.g. logo_v0p1.png) open Blender file, update textures, render at top, save as logo.png). Then make a copy of logo.png that is 640 x 360 called logo_small.png.
 
-- Change the version in `pyproject.toml`. Rather than X.X.X, use X.X.X.dev0. This version is temporary until you're positive that things are working properly.
+## 2. Create a git tag
 
-## 2. Build the distribution
+Versioning is handled dynamically based on git tags. So to make a new release, you first make a new git tag. Make sure you're on the main branch with no uncommitted changes.
+
+```bash
+RELEASE_VERSION=0.1.0
+git tag -a v${RELEASE_VERSION} -m "Release version ${RELEASE_VERSION}"
+git push origin v${RELEASE_VERSION}
+```
+
+If you need to delete a tag you've created (or even pushed), use:
+
+```bash
+git tag -d <tagname> # Locally delete
+git push origin :refs/tags/<tagname> # If you pushed it
+```
+
+## 3. Build the distribution
 
 ```bash
 poetry build
 ```
 
-You can find it at `dist/pooltool-0.3.3+dev.tar.gz` (along with the wheel (`.whl`))
+You should see something like this:
 
-## 3. Publish to the **test** PyPI repository
+```bash
+Building pooltool-billiards (0.1.0)
+  - Building sdist
+  - Built pooltool_billiards-0.1.0.tar.gz
+  - Building wheel
+  - Built pooltool_billiards-0.1.0-py3-none-any.whl
+```
+
+If there are additional metadata attached to the version (_e.g._ `0.3.4a2.dev1+eb17e9c.dirty`), then your tag isn't up to date.
+
+Open the tar found in `dist/`. If you've added any non-Python files to the package since the last release, make sure they are either present or absent from the package (depending on the desired outcome).
+
+Also open `pooltool/__init__.py` and make sure the `__version__` variable was populated with something other than the placeholder, `0.0.0`.
+
+## 4. Publish to the **test** PyPI repository
 
 - Populate your `.env` using `.env.copy` as a template.
 
@@ -40,24 +65,16 @@ conda activate asdf
 - Install:
 
 ```bash
-pip install --index-url https://test.pypi.org/simple/ --extra-index-url https://pypi.org/simple pooltool-billiards==X.X.X.dev0 --force-reinstall
+pip install --index-url https://test.pypi.org/simple/ --extra-index-url https://pypi.org/simple pooltool-billiards==${RELEASE_VERSION} --force-reinstall
 ```
 
 If test PyPi is timing out, you can instead consider testing the installation with 
 
 ```bash
-pip install dist/pooltool-billiards-X.X.X.dev0.tar.gz --force-reinstall
+pip install dist/pooltool-billiards-${RELEASE_VERSION}.tar.gz --force-reinstall
 ```
 
-- Test it out. Make sure `which run-pooltool` leads to the asdf environment: `/Users/evan/anaconda3/envs/asdf/bin/run_pooltool`. Then see if it works: `run-pooltool`. Additionally, check the path of `cd ~; python -c "import pooltool; print(pooltool.__file__)"; cd -`. It should be in site-packages of asdf environment.
-
-## 4. Rebuild with correct version tag
-
-- Change version to X.X.X in `pyproject.toml`, then **back in the development environment** create the dist:
-
-```bash
[…]
