### 2024-07-22 4d931ad
Move PYPI.md into docs
- It's content is not published on readthedocs
