### 2024-07-22 59dfc46
Add Makefile for publishing
