### 2024-07-26 b00b1c6
Platform-dependent panda3d install
