+++ pyproject.toml
@@ -82,13 +82,17 @@ priority = "explicit"
 
 [tool.poetry.dependencies]
 python = ">=3.8,<4.0"
+panda3d = [
+    {platform = "darwin", version=">=1.10.13,<1.11"},
+    {platform = "linux", version = "1.11.0.dev3005", allow-prereleases = true, source = "panda3d-archive"},
+    {platform = "win32", version = "1.11.0.dev3005", allow-prereleases = true, source = "panda3d-archive"},
+]
+panda3d-gltf = "*"
+panda3d-simplepbr = "*"
 numpy = "*"
 pandas = "*"
 numba = "*"
 scipy = "*"
-panda3d = { version = "1.11.0.dev3005", allow-prereleases = true, source = "panda3d-archive" }
-panda3d-gltf = "*"
-panda3d-simplepbr = "*"
 Pillow = "*"
 cattrs = "*"
 attrs = "*"
