### 2024-07-26 f976881
Drop pprofile for 3.12 support
- pprofile can be patched in if/when this merges:
  https://github.com/vpelletier/pprofile/pull/51
