+++ pyproject.toml
@@ -96,7 +96,6 @@ scipy = "*"
 Pillow = "*"
 cattrs = "*"
 attrs = "*"
-pprofile = "*"
 msgpack = "*"
 msgpack-numpy = "*"
 h5py = "*"
