### 2024-07-26 4bc962d
Move main.py into package
