### 2024-07-26 709fbb8
Glob the required assets during package build
