### 2024-07-28 74553f1
Scipy appears unused
