+++ pyproject.toml
@@ -93,7 +93,6 @@ panda3d-simplepbr = "*"
 numpy = "*"
 pandas = "*"
 numba = "*"
-scipy = "*"
 Pillow = "*"
 cattrs = "*"
 attrs = "*"
