+++ pooltool/utils/__init__.py
@@ -6,7 +6,6 @@ import os
 import pickle
 import tracemalloc
 
-import pandas as pd
 from panda3d.core import Filename
 
 
@@ -141,73 +140,6 @@ def display_top_memory_usage(snapshot, key_type="lineno", limit=10):
         print("%s other: %.1f KiB" % (len(other), size / 1024))
     total = sum(stat.size for stat in top_stats)
     print("Total allocated size: %.1f KiB" % (total / 1024))
-
-
-def memory_usage_to_dataframe(snapshot, key_type="lineno", limit=10):
-    """
-    Convert memory usage data from tracemalloc into a pandas DataFrame.
-
-    Args:
-    snapshot (tracemalloc.Snapshot): The snapshot of memory allocation.
-    key_type (str): The type of key to categorize memory usage.
-    limit (int): The number of top entries to include.
-
-    Returns:
-    pd.DataFrame: DataFrame containing memory usage information.
-    """
-
-    snapshot = snapshot.filter_traces(
-        (
-            tracemalloc.Filter(False, "<frozen importlib._bootstrap>"),
-            tracemalloc.Filter(False, "<frozen importlib._bootstrap_external>"),
-            tracemalloc.Filter(False, "<unknown>"),
-        )
-    )
-    top_stats = snapshot.statistics(key_type)
-
-    # Initialize lists to store data
-    ranks = []
-    files = []
-    line_numbers = []
-    memory_usages = []
-    code_snippets = []
-    categories = []
-
-    # Process top statistics
-    for index, stat in enumerate(top_stats[:limit], 1):
-        frame = stat.traceback[0]
-        filename = os.sep.join(frame.filename.split(os.sep)[-2:])
-        line = linecache.getline(frame.filename, frame.lineno).strip()
-
-        ranks.append(index)
-        files.append(filename)
-        line_numbers.append(frame.lineno)
-        memory_usages.append(stat.size / 1024)
-        code_snippets.append(line)
-        categories.append("Top")
-
-    # Process other statistics
-    other = top_stats[limit:]
-    if other:
-        size = sum(stat.size for stat in other)
-        ranks.append(None)
-        files.append(None)
-        line_numbers.append(None)
-        memory_usages.append(size / 1024)
-        code_snippets.append(None)
-        categories.append("Other")
-
-    # Create DataFrame
-    data = {
-        "Rank": ranks,
-        "File": files,
-        "Line Number": line_numbers,
-        "Memory Usage (KiB)": memory_usages,
-        "Code Snippet": code_snippets,
-        "Category": categories,
-    }
-
-    return pd.DataFrame(data)
 
 
 def human_readable_file_size(nbytes):
