+++ pyproject.toml
@@ -91,7 +91,6 @@ panda3d = [
 panda3d-gltf = "*"
 panda3d-simplepbr = "*"
 numpy = "*"
-pandas = "*"
 numba = "*"
 Pillow = "*"
 cattrs = "*"
