### 2024-07-28 57afd75
Switch dev version to 3.12
