### 2024-07-28 766896a
Drop 3.8 support
- Sensible pins for lower bounds of all dependencies
