+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
 build-backend = "poetry_dynamic_versioning.backend"
 
 [tool.poetry-dynamic-versioning]
-enable = true
+enable = false
 strict = true
 bump = true
 vcs = "git"
@@ -82,34 +82,38 @@ url = "https://archive.panda3d.org/simple/"
 priority = "explicit"
 
 [tool.poetry.dependencies]
-python = ">=3.8,<4.0"
+python = ">=3.9,<3.13"
 panda3d = [
-    {platform = "darwin", version=">=1.10.13,<1.11"},
-    {platform = "linux", version = "1.11.0.dev3005", allow-prereleases = true, source = "panda3d-archive"},
-    {platform = "win32", version = "1.11.0.dev3005", allow-prereleases = true, source = "panda3d-archive"},
+  {platform = "darwin", version=">=1.10.13,<1.11"},
+  {platform = "linux", version = "1.11.0.dev3005", allow-prereleases = true, source = "panda3d-archive"},
+  {platform = "win32", version = "1.11.0.dev3005", allow-prereleases = true, source = "panda3d-archive"},
 ]
-panda3d-gltf = "*"
-panda3d-simplepbr = "*"
-numpy = "*"
-numba = "*"
-Pillow = "*"
-cattrs = "*"
-attrs = "*"
-msgpack = "*"
-msgpack-numpy = "*"
-h5py = "*"
-pyyaml = "*"
-click = "^8.1.7"
+panda3d-gltf = ">=1.2.0"
+panda3d-simplepbr = ">=0.12.0"
+numpy = ">=1.26.0"  # Lower bound for 3.12 (https://github.com/numpy/numpy/releases/tag/v1.26.0)
+numba = ">=0.59.0"  # # Lower bound for 3.12 (https://numba.readthedocs.io/en/latest/user/installing.html#version-support-information)
+scipy = ">=1.12.0"  # Required for numba. Lower bound for 3.12 is officially 1.11, but in practice seems to be 1.12 on MacOS
+attrs = ">=21.3.0"
+cattrs = ">=22.1.0"
+msgpack = ">=1.0.0"  # cattrs structuring fails with msgpack<1
+msgpack-numpy = ">=0.4.8"
+pyyaml = ">=5.2"
+click = ">=8.0.0"
+Pillow = ">=6.2.0"
+h5py = ">=3.10"
 
 [tool.poetry.group.dev.dependencies]
-psutil = "*"
-colored = "*"
-ipython = "*"
-pytest = "*"
-pre-commit = "*"
-pyright = "*"
-ruff = "*"
-poetry-dynamic-versioning = {extras = ["plugin"], version = "^1.4.0"}
+# Testing and formatting
+pytest = ">=8.3.2"
+pre-commit = ">=3.7.1"
+pyright = ">=1.1.373"
+ruff = ">=0.5.5"
+# Diagnostics
+psutil = ">=6.0.0"
+colored = ">=2.2.4"  # TODO switch to something ubiquitous
+ipython = ">=8.18.1"
+# Publishing
+poetry-dynamic-versioning = {extras = ["plugin"], version = ">=1.4.0"}
 
 [tool.poetry.group.docs]
 optional = true
