### 2024-07-28 e0e3cbb
Run pytest on lower + upper supported py versions
