### 2024-07-28 8c29ba7
Python 3.12 compatible panda3d 1.11 version
