+++ .github/workflows/lint-and-check.yml
@@ -1,42 +1,32 @@
-name: lint-and-check
+name: format
 
 on:
   pull_request:
     branches: [main]
 
 jobs:
-  lint-and-check:
+  format:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v3
-
-      - name: Set up Python
-        uses: actions/setup-python@v4
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v4
         with:
-          python-version: 3.12
-
-      - name: Cache Poetry install
-        uses: actions/cache@v2
+          python-version: "3.12"
+      - name: Install poetry
+        uses: abatilo/actions-poetry@v2
         with:
-          path: ~/.local
-          key: poetry-1.8.3-0
-  
-      - name: Install Poetry
-        uses: snok/install-poetry@v1
+          poetry-version: "1.8.3"
+      - name: Setup a local virtual environment
+        run: |
+          poetry config virtualenvs.create true --local
+          poetry config virtualenvs.in-project true --local
+      - uses: actions/cache@v3
+        name: Define a cache for the virtual environment based on the dependencies lock file
         with:
-          version: 1.8.3
-          virtualenvs-create: true
-          virtualenvs-in-project: true
-  
-      - name: Cache Poetry dependencies
-        id: cache-poetry-deps
-        uses: actions/cache@v2
-        with:
-          path: .venv
-          key: pydeps-${{ hashFiles('**/poetry.lock') }}
-
-      - run: poetry install --no-interaction --no-root
-        if: steps.cache-poetry-deps.outputs.cache-hit != 'true'
+          path: ./.venv
+          key: venv-${{ hashFiles('poetry.lock') }}
+      - name: Install project dependencies
+        run: poetry install
 
       # --- Ruff
 
@@ -68,7 +58,7 @@ jobs:
 
       # --- Main
 
-      - name: Test results
+      - name: Results
         if: always()
         run: |
           # Print out test results
@@ -108,7 +98,7 @@ jobs:
               echo "    - $check"
             done
           else
-            echo "🚀🚀 ALL TESTS PASSED 🚀🚀"
+            echo "🚀🚀 ALL FORMATTING & TYPE-CHECKING PASSED 🚀🚀"
           fi
 
           echo ""
