+++ .github/workflows/test.yml
@@ -5,60 +5,33 @@ on:
     branches: [main]
 
 jobs:
-  pytest:
-    runs-on: ubuntu-latest
+  test:
     strategy:
+      fail-fast: false
       matrix:
-        python-version: [3.9, 3.12]  # Define the Python versions you want to test
-
+        python-version: ["3.9", "3.12"]
+        poetry-version: ["1.8.3"]
+        os: [ubuntu-latest, macos-latest, windows-latest]
+    runs-on: ${{ matrix.os }}
     steps:
-      - uses: actions/checkout@v3
-
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v4
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
-
-      - name: Cache Poetry install
-        uses: actions/cache@v2
+      - name: Install poetry
+        uses: abatilo/actions-poetry@v2
         with:
-          path: ~/.local
-          key: poetry-1.8.3-${{ matrix.python-version }}-0
-  
-      - name: Install Poetry
-        uses: snok/install-poetry@v1
-        with:
-          version: 1.8.3
-          virtualenvs-create: true
-          virtualenvs-in-project: true
-  
-      - name: Cache Poetry dependencies
-        id: cache-poetry-deps
-        uses: actions/cache@v2
-        with:
-          path: .venv
-          key: pydeps-${{ matrix.python-version }}-${{ hashFiles('**/poetry.lock') }}
-
-      - run: poetry install --no-interaction --no-root
-        if: steps.cache-poetry-deps.outputs.cache-hit != 'true'
-
-      # --- Pytest
-
-      - name: pytest
-        id: pytest
-        continue-on-error: true
+          poetry-version: ${{ matrix.poetry-version }}
+      - name: Setup a local virtual environment
         run: |
-          poetry run pytest --version
-          poetry run pytest
-          echo "pytest_failed=$?" >> $GITHUB_ENV
-
-      - name: Test results
-        if: always()
-        run: |
-          # Print out test results
-          if [[ "${{ env.pytest_failed }}" != "0" ]]; then
-            echo "❌ pytest FAILED"
-            exit 1
-          else
-            echo "✅ pytest PASSED"
-          fi
+          poetry config virtualenvs.create true --local
+          poetry config virtualenvs.in-project true --local
+      - uses: actions/cache@v3
+        name: Define a cache for the virtual environment based on the dependencies lock file
+        with:
+          path: ./.venv
+          key: venv-${{ hashFiles('poetry.lock') }}
+      - name: Install project dependencies
+        run: poetry install
+      - name: Run tests
+        run: poetry run pytest
