### 2024-07-28 7aac9a3
Update pip install instructions
