### 2024-07-28 10e299e
Enable dynamic versioning
