+++ docs/publish_instructions.md
@@ -88,7 +88,7 @@ pip install dist/pooltool_billiards-${RELEASE_VERSION}.tar.gz --force-reinstall 
 source ~/.bashrc
 conda deactivate
 conda env remove --name asdf
-conda create -y -n asdf python=3.8.10
+conda create -y -n asdf python=3.12.4
 PYTHONPATH=""
 conda activate asdf
 cd ~
