+++ pyproject.toml
@@ -101,6 +101,7 @@ pyyaml = ">=5.2"
 click = ">=8.0.0"
 Pillow = ">=6.2.0"
 h5py = ">=3.10"
+pytest-cov = "^5.0.0"
 
 [tool.poetry.group.dev.dependencies]
 # Testing and formatting
