### 2024-08-03 c915580
Add project status badge
