### 2024-08-03 d1410e5
Add latest PyPI version badge
