### 2024-08-03 fb16db9
Add code coverage badge
