### 2024-08-03 f062fba
Move pytest cov to "test" group dependency
