### 2024-08-03 a11b64c
Fix pyright error
