+++ .github/workflows/test.yml
@@ -33,12 +33,15 @@ jobs:
           key: venv-${{ hashFiles('poetry.lock') }}
       - name: Install project dependencies
         run: poetry install
-      - name: Run tests and build coverage statistics
-        run: poetry run pytest --cov pooltool ${{ env.CODECOV_ATS_TESTS }}
-      - name: Upload test coverage to Codecov
+      - name: Run tests
+        run: poetry run pytest
+
+      # Calculate the test coverage and upload results
+      - name: Get and upload test coverage
         if: matrix.python-version == '3.12' && matrix.os == 'ubuntu-latest'
         run: |
           # Replace `linux` below with the appropriate OS (Options are `alpine`, `linux`, `macos`, `windows`)
+          poetry run pytest --cov pooltool --cov-report=xml
           curl -Os https://cli.codecov.io/latest/linux/codecov
           chmod +x codecov
-          ./codecov --verbose upload-process --fail-on-error -t ${{ secrets.CODECOV_TOKEN }} -n 'service'-${{ github.run_id }} -F service -f coverage-service.xml
+          ./codecov --verbose upload-process --fail-on-error -t ${{ secrets.CODECOV_TOKEN }} -n 'service'-${{ github.run_id }} -F service -f coverage.xml
