### 2024-08-04 af745d0
Reorganize badges
