### 2024-08-04 b24ba65
Add CI badge
