### 2024-08-04 6ea45ef
Add sphinx-tabs as dependency
