### 2024-08-04 044fec6
Streamline installation instruction formatting
