+++ docs/getting_started/install.md
@@ -1,58 +1,48 @@
 # Installation
 
-You're just a few seconds away from exploring the interactive interface.
+Pooltool is available on Linux, Mac, and Windows for the following Python versions:
 
-```{figure} ../_assets/gallery_6.jpg
----
-height: 300px
----
+![PyPI - Python Version](https://img.shields.io/pypi/pyversions/pooltool-billiards)
+
+## With pip:
+
+Pooltool is hosted on the [Python Package Index (PyPI)](https://pypi.org/project/pooltool-billiards/) and can be installed with pip.
+
+```{eval-rst}
+.. tabs::
+
+   .. tab:: Linux
+
+      .. code-block:: bash
+
+         pip install pooltool-billiards --extra-index-url https://archive.panda3d.org/
+
+      (*Providing the Panda3D archive is required until Panda3D v1.11 is released*)
+
+   .. tab:: MacOS
+
+      .. code-block:: bash
+
+         pip install pooltool-billiards
+
+   .. tab:: Windows
+
+      .. code-block:: bash
+
+         pip install pooltool-billiards --extra-index-url https://archive.panda3d.org/
+
+      (*Providing the Panda3D archive is required until Panda3D v1.11 is released*)
 ```
 
-```{figure} ../_assets/gallery_7.jpg
----
-height: 300px
----
-```
-
-## Requirements
-
-Pooltool is compatible with Python >=3.9,<3.13.
-
-## Install option (1): pip
-
-| GUI | API | Develop |
-|-----|-----|---------|
-| ✅  | ✅  | ❌      |
-
-<details><summary style="font-style: italic;">[Click to expand/collapse]</summary>
-
-### MacOS
-
-```python
-pip install pooltool-billiards
-```
-
-### Linux & Windows
-
-```python
-pip install pooltool-billiards --extra-index-url https://archive.panda3d.org/
-```
-
-</details>
-
-## Install option (2): developer
-
-| GUI | API | Develop |
-|-----|-----|---------|
-| ✅  | ✅  | ✅      |
-
-<details><summary style="font-style: italic;">[Click to expand/collapse]</summary>
+## From source:
 
 If you want to develop for pooltool, have access to the most up-to-date version of the codebase, or modify the code to your liking, this is for you.
 
+<details><summary style="font-style: italic;">[Click to expand/collapse]</summary>
+
 A small note. If you don't have the ability to create isolated python environments, I would recommend installing `conda` ([here](https://conda.io/projects/conda/en/latest/user-guide/install/index.html)) so you can isolate pooltool from your other business.
 
-**(i)** Grab a copy of the codebase.
+**1.** Grab a copy of the codebase.
 
 ```bash
 cd <A_DIRECTORY_YOU_LIKE>
@@ -60,7 +50,7 @@ git clone https://github.com/ekiefl/pooltool.git
 cd pooltool
 ```
 
-**(ii)** Create a new python environment that uses Python 3.12.4.
+**2.** Create a new python environment that uses Python 3.12.4.
 
 If you have `conda`, just run this:
 
@@ -73,15 +63,14 @@ Regardless of how you managed your python environment, please verify you're runn
 
 ```bash
 $ python
-Python 3.8.10 (default, May 19 2021, 11:01:55)
-[Clang 10.0.0 ] :: Anaconda, Inc. on darwin
+Python 3.12.4 | packaged by Anaconda, Inc. | (main, Jun 18 2024, 10:14:12) [Clang 14.0.6 ] on darwin
 Type "help", "copyright", "credits" or "license" for more information.
 >>> exit()
 ```
 
-**(iii)** Install poetry, a popular python package/environment manager.
+**3.** Install poetry, a popular python package/environment manager.
 
-If you created your environment with conda, you've already installed poetry.
[…]
