+++ response.md
@@ -1,0 +1,226 @@
+# Response to reviewers (take one)
+
+Thanks again for your reviews @sebastianotronto and @eliotwrobson.
+
+## Common themes
+
+Reading through everything, there are a couple of major themes that I'm picking up on that I want to address at a high level. Those things are versioning, ease of installation, and lack of obvious research application. 
+
+### (1) Versioning
+
+Thanks both of you for pointing out that pooltool doesn't work on Python 3.12. The culprit was the dependency `pprofile`, which is used in a couple of places in `sandbox/`. I think it's a great package, but until https://github.com/vpelletier/pprofile/pull/51 is merged, I'm forced to drop it as a dependency. Any profiling code has been removed, rather than replaced with something else ¯\_(ツ)_/¯.
+
+Package dependencies are now defined by poetry. I did some manual dependency version testing and settled on the following dependencies:
+
+```
+[tool.poetry.dependencies]
+python = ">=3.9,<3.13"
+panda3d = [
+  {platform = "darwin", version=">=1.10.13,<1.11"},
+  {platform = "linux", version = "1.11.0.dev3444", allow-prereleases = true, source = "panda3d-archive"},
+  {platform = "win32", version = "1.11.0.dev3444", allow-prereleases = true, source = "panda3d-archive"},
+]
+panda3d-gltf = ">=1.2.0"
+panda3d-simplepbr = ">=0.12.0"
+numpy = ">=1.26.0"  # Lower bound for 3.12 (https://github.com/numpy/numpy/releases/tag/v1.26.0)
+numba = ">=0.59.0"  # # Lower bound for 3.12 (https://numba.readthedocs.io/en/latest/user/installing.html#version-support-information)
+scipy = ">=1.12.0"  # Required for numba. Lower bound for 3.12 is officially 1.11, but in practice seems to be 1.12 on MacOS
+attrs = ">=21.3.0"
+cattrs = ">=22.1.0"
+msgpack = ">=1.0.0"  # cattrs structuring fails with msgpack<1
+msgpack-numpy = ">=0.4.8"
+pyyaml = ">=5.2"
+click = ">=8.0.0"
+Pillow = ">=6.2.0"
+h5py = ">=3.10"
+```
+
+
+The lower bounds were determined judiciously, where I tried to pick the oldest versions that were compatible with Python 3.12. This maximizes the target space for dependency resolution so as to maximize the odds that pooltool can share the same python environment with other packages, rather than being relegated as a standalone application. At the same time, I didn't create any dependency branching based on Python version (e.g. if Python <3.9, use Numpy <1.26).
+
+Speaking of which, numpy 1.26.0 is the lower bound for 3.12, but it drops support for 3.8. I've followed suit and dropped support for 3.8, which makes pooltool's Python range >=3.9,<3.13.
+
+To test over this version range, I am now running pooltool's test suite in a matrix of builds, where I test 3.9 and 3.12 for Linux, MacOS, and Windows. With this test matrix now in place, testing for 3.13 and beyond will be easy-peasy.
+
+The development version has correspondingly experienced a _massive_ bump from 3.8.10 to 3.12.4. No more living in the past!
+
+All the changes mentioned can be found in https://github.com/ekiefl/pooltool/pull/124 and https://github.com/ekiefl/pooltool/pull/125.
+
+### (2) Ease of installation
+
+Installation got mentioned quite a few times, which obviously isn't a great sign. For me, the main takeaways from reading your comments were that installation (1) doesn't work on 3.12, (2) perhaps messy/long, and (3) somewhat hard to find.
+
+As mentioned, pooltool now works on 3.12. There also now exists improved signage in the [installation instructions](https://pooltool.readthedocs.io/en/latest/getting_started/install.html#installation) and in the README.md that clarify which Python versions pooltool supports.
+
[…]
