### 2024-08-06 #173 @eliotwrobson
@ekiefl As a heads up, I found some time to finish my review! It looks like you were able to address most of the items from my previous round of comments. The only things remaining from my end are the following:

1. I can see the code coverage badge but it doesn't show the coverage itself, maybe you just need to push again or something? I don't think this is really a huge deal.
2. I think that the JOSS submission is solid, but having some information about the usage of the package would be nice. Including some code from a vignette would be a good way to add this without too much extra effort (which ties into the third item), and maybe having some screenshots from the GUI would be cool too.
3. The other remaining checklist items are all related to vignettes for the package. I was able to get the game interface working reasonably well, and the discussion in the documentation was easy to follow, but I couldn't find vignettes for a simulation task that might arise in research not using the GUI. Having a couple of these would be awesome.
