### 2024-08-07 #173 @cmarmo
Thank you @ekiefl for your prompt answer and the todo summary!
I'm letting @sebastianotronto and @eliotwrobson comment about your todo list.

>  is right that not every user-facing function contains an example. But as they've pointed out, that's because the API is very granular: "As I noted above, not all functions have examples, but considering how fine-grained the API is I do not think it would be necessary".

We experienced similar situations with other consistent packages. As we try to enforce a review schedule, my suggestion here is to add some of the missing examples to the documentation, in order to signify the "work in progress" status, and let the rest for future improvements.... you may even want to organize some sprints and call for contributors in that case...

I have labeled this issue as "awaiting changes" : let's touch base in three weeks, at August 28th, to see if reviewers are happy with your changes.

Does this sound reasonable?
