### 2024-08-07 #173 @sebastianotronto
Hi @ekiefl, thank you for your response!

After your comment I checked the recent changes and ticked a few more checkboxes. Also, I would like to point out that some of the unchecked boxes are (or were) so because the requested information was not present *in the README file*. But I agree with you that having it in the docs and linking to it in the README is a good approach. @cmarmo can the requirements be considered satisfied if the information is only linked to in the README?

Thanks for improving the installation process, I tested it now and the one-liner works!

As for the research applications, I admit that at the time of my first review I did not understand that the project was also submitted to JOSS, hence the missing ticks. The goal of the project is clear to me and I understand the research applications, all good on that side!

For now I am very happy with the changes you made, I'll check back in a couple of weeks for the items in your to-do list :)
