### 2024-08-11 d6e5d96
Split up minimum_quartic_root implementation
