+++ pooltool/ptmath/roots/core.py
@@ -1,3 +1,5 @@
+from typing import Tuple
+
 import numpy as np
 from numba import jit
 from numpy.typing import NDArray
@@ -5,7 +7,7 @@ from numpy.typing import NDArray
 import pooltool.constants as const
 
 
-def min_real_root(
+def min_real_root_old(
     roots: NDArray[np.complex128],
     abs_or_rel_cutoff: float = 1e-3,
     rtol: float = 1e-3,
@@ -62,6 +64,92 @@ def min_real_root(
 
     # Return candidate with the smallest real component
     return candidates[candidates.real.argmin()]
+
+
+def min_real_root(
+    roots: NDArray[np.complex128],
+    abs_or_rel_cutoff: float = 1e-3,
+    rtol: float = 1e-3,
+    atol: float = 1e-9,
+) -> np.complex128:
+    """Find the minimum real positive root, using the processed roots from process_roots.
+
+    Args:
+        roots:
+            A 1D array of roots.
+        abs_or_rel_cutoff:
+            The criteria for a root being real depends on the magnitude of its real
+            component.
+        atol:
+            A root r (with abs(r.real) >= abs_or_rel_cutoff) is considered real if
+            abs(r.imag) < atol.
+        rtol:
+            A root r (with abs(r.real) < abs_or_rel_cutoff) is considered real if
+            abs(r.imag) / abs(r.real) < rtol.
+
+    Returns:
+        root:
+            The smallest, real, positive root. Returns np.inf if no such root exists.
+    """
+    # Process the roots to filter out non-real roots
+    roots, is_real = process_roots(roots, abs_or_rel_cutoff, rtol, atol)
+
+    # Select the smallest real root (which are not set to np.inf)
+    if roots.size == 0 or np.all(roots == np.inf):
+        return np.complex128(np.inf)
+
+    min_root = np.min(roots[is_real])
+    return np.complex128(min_root)
+
+
+def process_roots(
+    roots: NDArray[np.complex128],
+    abs_or_rel_cutoff: float = 1e-3,
+    rtol: float = 1e-3,
+    atol: float = 1e-9,
+) -> Tuple[NDArray[np.complex128], NDArray[np.bool]]:
+    """Process roots and return an array where non-real roots are set to inf.
+
+    Args:
+        roots:
+            A 1D array of roots.
+        abs_or_rel_cutoff:
+            The criteria for a root being real depends on the magnitude of its real
+            component. If it's large, we require the imaginary component to be less than
+            atol in absolute terms. But when the real component is small, we require the
+            imaginary component be less than a fraction, rtol, of the real component.
+            abs_or_rel_cutoff defines a threshold for the magnitude of the real
+            component, above which atol is used and below which rtol is used.
+        atol:
+            A root r (with abs(r.real) >= abs_or_rel_cutoff) is considered real if
+            abs(r.imag) < atol.
+        rtol:
+            A root r (with abs(r.real) < abs_or_rel_cutoff) is considered real if
+            abs(r.imag) / abs(r.real) < rtol. And in the special case when r.real == 0,
+            the root is considered real if r.imag == 0, too.
+
+    Returns:
+        (NDArray[np.complex128], NDArray[np.bool]):
+            - roots: The processed roots where non-real roots are set to infinity.
+            - is_real: A boolean array indicating which roots are considered real.
+    """
+    positive = roots.real >= 0.0
+
+    imag_mag = np.abs(roots.imag)
+    real_mag = np.abs(roots.real)
+
+    big = real_mag > abs_or_rel_cutoff
+    big_keep = (imag_mag < atol) & positive
+
[…]
