### 2024-08-12 #173 @eliotwrobson
Still reading but all of your proposed solutions for issue 3 sound really good 👍🏽 Mentioning something about these in your JOSS paper would make it stronger as well.
