### 2024-08-12 #173 @cmarmo
> can the requirements be considered satisfied if the information is only linked to in the README?

@sebastianotronto sure! The important thing is that the information is accessible from the README: links are an acceptable option.
