### 2024-08-18 b160556
Refactor of quartics
