### 2024-08-19 418f951
Working yet untested ball-ball caching
