+++ pooltool/evolution/event_based/simulate.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 from itertools import combinations
-from typing import Dict, Optional, Set, Tuple
+from typing import Dict, List, Optional, Set, Tuple, cast
 
 import attrs
 import numpy as np
@@ -152,6 +152,7 @@ def simulate(
         engine.resolver.resolve(shot, event)
         shot._update_history(event)
 
+    collision_cache = CollisionCache()
     transition_cache = TransitionCache.create(shot)
 
     events = 0
@@ -159,6 +160,7 @@ def simulate(
         event = get_next_event(
             shot,
             transition_cache=transition_cache,
+            collision_cache=collision_cache,
             quartic_solver=quartic_solver,
         )
 
@@ -171,6 +173,7 @@ def simulate(
         if event.event_type in include:
             engine.resolver.resolve(shot, event)
             transition_cache.update(event)
+            collision_cache.invalidate(event)
 
         shot._update_history(event)
 
@@ -318,22 +321,77 @@ def _next_transition(ball: Ball) -> Event:
 
 @attrs.define
 class CollisionCache:
-    ball_ball: Dict[Tuple[str, str], float] = attrs.field(factory=dict)
+    times: Dict[EventType, Dict[Tuple[str, str], float]] = attrs.field(factory=dict)
+
+    @property
+    def size(self) -> int:
+        count = 0
+        for cache in self.times.values():
+            count += len(cache)
+
+        return count
+
+    def _get_invalid_ball_ids(self, event: Event) -> Set[str]:
+        invalid_ball_ids = set()
+
+        if event.event_type == EventType.BALL_BALL:
+            invalid_ball_ids.update(event.ids)
+        elif event.event_type in {
+            EventType.BALL_LINEAR_CUSHION,
+            EventType.BALL_CIRCULAR_CUSHION,
+            EventType.BALL_POCKET,
+        }:
+            invalid_ball_ids.add(event.agents[0].id)
+        elif event.event_type == EventType.STICK_BALL:
+            invalid_ball_ids.add(event.agents[1].id)
+        else:
+            assert event.event_type.is_transition()
+            invalid_ball_ids.add(event.agents[0].id)
+
+        return invalid_ball_ids
+
+    def invalidate(self, event: Event) -> None:
+        invalid_ball_ids = self._get_invalid_ball_ids(event)
+
+        for event_type, event_times in self.times.items():
+            keys_to_delete = []
+
+            for key in event_times:
+                if event_type == EventType.BALL_BALL:
+                    if key[0] in invalid_ball_ids or key[1] in invalid_ball_ids:
+                        keys_to_delete.append(key)
+                elif event_type in {
+                    EventType.BALL_LINEAR_CUSHION,
+                    EventType.BALL_CIRCULAR_CUSHION,
+                    EventType.BALL_POCKET,
+                }:
+                    if key[0] in invalid_ball_ids:
+                        keys_to_delete.append(key)
+                elif event_type == EventType.STICK_BALL:
+                    if key[1] in invalid_ball_ids:
+                        keys_to_delete.append(key)
+
+            for key in keys_to_delete:
+                del event_times[key]
 
 
 def get_next_ball_ball_collision(
     shot: System,
-    collision_cache: Optional[CollisionCache] = None,
+    collision_cache: CollisionCache,
     solver: QuarticSolver = QuarticSolver.HYBRID,
 ) -> Event:
     """Returns next ball-ball collision"""
 
-    dtau_E = np.inf
+    ball_pairs: List[Tuple[str, str]] = []
+    collision_coeffs: List[Tuple[float, ...]] = []
 
-    ball_pairs = []
-    collision_coeffs = []
[…]
