+++ pooltool/evolution/event_based/simulate.py
@@ -3,16 +3,14 @@
 from __future__ import annotations
 
 from itertools import combinations
-from typing import Dict, List, Optional, Set, Tuple, cast
+from typing import List, Optional, Set, Tuple, cast
 
-import attrs
 import numpy as np
 
 import pooltool.constants as const
 import pooltool.physics.evolve as evolve
 import pooltool.ptmath as ptmath
 from pooltool.events import (
-    AgentType,
     Event,
     EventType,
     ball_ball_collision,
@@ -20,14 +18,11 @@ from pooltool.events import (
     ball_linear_cushion_collision,
     ball_pocket_collision,
     null_event,
-    rolling_spinning_transition,
-    rolling_stationary_transition,
-    sliding_rolling_transition,
-    spinning_stationary_transition,
     stick_ball_collision,
 )
 from pooltool.evolution.continuize import continuize
 from pooltool.evolution.event_based import solve
+from pooltool.evolution.event_based.cache import CollisionCache, TransitionCache
 from pooltool.evolution.event_based.config import INCLUDED_EVENTS
 from pooltool.objects.ball.datatypes import Ball, BallState
 from pooltool.objects.table.components import (
@@ -40,6 +35,29 @@ from pooltool.ptmath.roots.quartic import QuarticSolver, solve_quartics
 from pooltool.system.datatypes import System
 
 DEFAULT_ENGINE = PhysicsEngine()
+
+
+def _evolve(shot: System, dt: float):
+    """Evolves current ball an amount of time dt
+
+    FIXME This is very inefficent. each ball should store its natural trajectory
+    thereby avoid a call to the clunky evolve_ball_motion. It could even be a
+    partial function so parameters don't continuously need to be passed
+    """
+
+    for ball in shot.balls.values():
+        rvw, _ = evolve.evolve_ball_motion(
+            state=ball.state.s,
+            rvw=ball.state.rvw,
+            R=ball.params.R,
+            m=ball.params.m,
+            u_s=ball.params.u_s,
+            u_sp=ball.params.u_sp,
+            u_r=ball.params.u_r,
+            g=ball.params.g,
+            t=dt,
+        )
+        ball.state = BallState(rvw, ball.state.s, shot.t + dt)
 
 
 def simulate(
@@ -193,29 +211,6 @@ def simulate(
     return shot
 
 
-def _evolve(shot: System, dt: float):
-    """Evolves current ball an amount of time dt
-
-    FIXME This is very inefficent. each ball should store its natural trajectory
-    thereby avoid a call to the clunky evolve_ball_motion. It could even be a
-    partial function so parameters don't continuously need to be passed
-    """
-
-    for ball in shot.balls.values():
-        rvw, _ = evolve.evolve_ball_motion(
-            state=ball.state.s,
-            rvw=ball.state.rvw,
-            R=ball.params.R,
-            m=ball.params.m,
-            u_s=ball.params.u_s,
-            u_sp=ball.params.u_sp,
-            u_r=ball.params.u_r,
-            g=ball.params.g,
-            t=dt,
-        )
-        ball.state = BallState(rvw, ball.state.s, shot.t + dt)
-
-
 def get_next_event(
     shot: System,
     *,
@@ -257,122 +252,6 @@ def get_next_event(
         event = ball_pocket_event
 
     return event
-
-
-def _null() -> Dict[str, Event]:
-    return {"null": null_event(time=np.inf)}
-
-
-@attrs.define
-class TransitionCache:
-    transitions: Dict[str, Event] = attrs.field(factory=_null)
-
-    def get_next(self) -> Event:
-        return min(
-            (trans for trans in self.transitions.values()), key=lambda event: event.time
-        )
[…]
