### 2024-08-20 f120f85
Simplify logic
