### 2024-08-20 0ad4acb
Add ball-pocket collision cache
