+++ pooltool/evolution/event_based/simulate.py
@@ -24,11 +24,7 @@ from pooltool.evolution.continuize import continuize
 from pooltool.evolution.event_based import solve
 from pooltool.evolution.event_based.cache import CollisionCache, TransitionCache
 from pooltool.evolution.event_based.config import INCLUDED_EVENTS
-from pooltool.objects.ball.datatypes import Ball, BallState
-from pooltool.objects.table.components import (
-    CircularCushionSegment,
-    Pocket,
-)
+from pooltool.objects.ball.datatypes import BallState
 from pooltool.physics.engine import PhysicsEngine
 from pooltool.ptmath.roots.quartic import QuarticSolver, solve_quartics
 from pooltool.system.datatypes import System
@@ -236,19 +232,21 @@ def get_next_event(
     if ball_ball_event.time < event.time:
         event = ball_ball_event
 
+    ball_circular_cushion_event = get_next_ball_circular_cushion_event(
+        shot, collision_cache=collision_cache, solver=quartic_solver
+    )
+    if ball_circular_cushion_event.time < event.time:
+        event = ball_circular_cushion_event
+
     ball_linear_cushion_event = get_next_ball_linear_cushion_collision(
         shot, collision_cache=collision_cache
     )
     if ball_linear_cushion_event.time < event.time:
         event = ball_linear_cushion_event
 
-    ball_circular_cushion_event = get_next_ball_circular_cushion_event(
-        shot, solver=quartic_solver
+    ball_pocket_event = get_next_ball_pocket_collision(
+        shot, collision_cache=collision_cache, solver=quartic_solver
     )
-    if ball_circular_cushion_event.time < event.time:
-        event = ball_circular_cushion_event
-
-    ball_pocket_event = get_next_ball_pocket_collision(shot, solver=quartic_solver)
     if ball_pocket_event.time < event.time:
         event = ball_pocket_event
 
@@ -347,22 +345,35 @@ def get_next_ball_ball_collision(
 
 
 def get_next_ball_circular_cushion_event(
-    shot: System, solver: QuarticSolver = QuarticSolver.HYBRID
+    shot: System,
+    collision_cache: CollisionCache,
+    solver: QuarticSolver = QuarticSolver.HYBRID,
 ) -> Event:
     """Returns next ball-cushion collision (circular cushion segment)"""
 
-    dtau_E = np.inf
-    agent_ids = []
-    collision_coeffs = []
+    if not shot.table.has_circular_cushions:
+        return null_event(np.inf)
+
+    ball_cushion_pairs: List[Tuple[str, str]] = []
+    collision_coeffs: List[Tuple[float, ...]] = []
+
+    cache = collision_cache.times.setdefault(EventType.BALL_CIRCULAR_CUSHION, {})
 
     for ball in shot.balls.values():
-        if ball.state.s in const.nontranslating:
-            continue
-
         state = ball.state
         params = ball.params
 
         for cushion in shot.table.cushion_segments.circular.values():
+            obj_ids = (ball.id, cushion.id)
+
+            if obj_ids in cache:
+                continue
+
+            if ball.state.s in const.nontranslating:
+                cache[obj_ids] = np.inf
+                continue
+
+            ball_cushion_pairs.append(obj_ids)
             collision_coeffs.append(
                 solve.ball_circular_cushion_collision_coeffs(
                     rvw=state.rvw,
@@ -377,30 +388,29 @@ def get_next_ball_circular_cushion_event(
                 )
             )
 
-            agent_ids.append((ball.id, cushion.id))
+    if len(collision_coeffs):
[…]
