### 2024-08-24 6f7d1dd
Remove asserts and casts
