### 2024-08-24 f39c9fd
Fix null event conditional
