### 2024-08-24 5ee0a32
strict=True doesn't exist in Python 3.9
