### 2024-08-24 7eb2923
Remove multiline call signatures
