+++ docs/conf.py
@@ -173,5 +173,4 @@ rst_prolog = """
 # Related custom CSS
 html_css_files = [
     "css/label.css",
-    "css/sig.css",
 ]
