### 2024-08-24 1c03d13
Move tests to tests/, not working
