+++ pyproject.toml
@@ -135,6 +135,9 @@ line-length = 88
 extend-select = ["I"]
 
 [tool.pytest.ini_options]
+addopts = [
+    "--import-mode=importlib",
+]
 testpaths = [
     "tests",
 ]
