### 2024-08-25 99217e5
Add types for numba
