+++ pooltool/ptmath/utils.py
@@ -1,5 +1,5 @@
 from math import degrees, sqrt
-from typing import Tuple
+from typing import Callable, Tuple
 
 import numpy as np
 from numba import jit
@@ -8,7 +8,13 @@ from numpy.typing import NDArray
 import pooltool.constants as const
 
 
-def solve_transcendental(f, a, b, tol=1e-5, max_iter=100) -> float:
+def solve_transcendental(
+    f: Callable[[float], float],
+    a: float,
+    b: float,
+    tol: float = 1e-5,
+    max_iter: int = 100,
+) -> float:
     """Solve transcendental equation f(x) = 0 in interval [a, b] using bisection method
 
     Args:
@@ -50,50 +56,23 @@ def solve_transcendental(f, a, b, tol=1e-5, max_iter=100) -> float:
     return c
 
 
-@jit(nopython=True, cache=const.use_numba_cache)
-def orientation(p, q, r):
-    """Find the orientation of an ordered triplet (p, q, r)
-
-    See https://www.geeksforgeeks.org/orientation-3-ordered-points/amp/
-
-    Notes
-    =====
-    - 3D points may be passed but only the x and y components are used
-
-    Returns
-    =======
-    output : int
-        0 : Collinear points, 1 : Clockwise points, 2 : Counterclockwise
-    """
-    val = ((q[1] - p[1]) * (r[0] - q[0])) - ((q[0] - p[0]) * (r[1] - q[1]))
-    if val > 0:
-        # Clockwise orientation
-        return 1
-    elif val < 0:
-        # Counterclockwise orientation
-        return 2
-    else:
-        # Collinear orientation
-        return 0
-
-
-def convert_2D_to_3D(array: NDArray) -> NDArray:
+def convert_2D_to_3D(array: NDArray[np.float64]) -> NDArray[np.float64]:
     """Convert a 2D vector to a 3D vector, setting z=0"""
-    return np.pad(array, (0, 1), "constant", constant_values=(0,))  # type: ignore
+    return np.pad(array, (0, 1), "constant", constant_values=(0,))
 
 
-def angle_between_vectors(v1, v2) -> float:
+def angle_between_vectors(v1: NDArray[np.float64], v2: NDArray[np.float64]) -> float:
     """Returns angles between [-180, 180]"""
     angle = np.math.atan2(np.linalg.det([v1, v2]), np.dot(v1, v2))  # type: ignore
     return degrees(angle)
 
 
-def wiggle(x, val):
+def wiggle(x: float, val: float):
     """Vary a float or int x by +- val according to a uniform distribution"""
     return x + val * (2 * np.random.rand() - 1)
 
 
-def are_points_on_same_side(p1, p2, p3, p4):
+def are_points_on_same_side(p1, p2, p3, p4) -> bool:
     """Are points p3, p4 are on the same side of the line formed by points p1 and p2?
 
     Accepts indexable objects. This is a 2D function, but if higher dimensions are
@@ -147,7 +126,7 @@ def find_intersection_2D(
 
 
 @jit(nopython=True, cache=const.use_numba_cache)
-def cross(u, v):
+def cross(u: NDArray[np.float64], v: NDArray[np.float64]) -> NDArray[np.float64]:
     """Compute cross product u x v, where u and v are 3-dimensional vectors
 
     (just-in-time compiled)
@@ -161,7 +140,9 @@ def cross(u, v):
     )
 
 
-def unit_vector_slow(vector, handle_zero=False):
+def unit_vector_slow(
[…]
