+++ pooltool/evolution/event_based/solve.py
@@ -1,7 +1,9 @@
 from math import acos
+from typing import Tuple
 
 import numpy as np
 from numba import jit
+from numpy.typing import NDArray
 
 import pooltool.constants as const
 import pooltool.physics.evolve as evolve
@@ -9,7 +11,14 @@ import pooltool.ptmath as ptmath
 
 
 @jit(nopython=True, cache=const.use_numba_cache)
-def skip_ball_ball_collision(rvw1, rvw2, s1, s2, R1, R2):
+def skip_ball_ball_collision(
+    rvw1: NDArray[np.float64],
+    rvw2: NDArray[np.float64],
+    s1: int,
+    s2: int,
+    R1: float,
+    R2: float,
+) -> bool:
     if (s1 == const.spinning or s1 == const.pocketed or s1 == const.stationary) and (
         s2 == const.spinning or s2 == const.pocketed or s2 == const.stationary
     ):
@@ -71,7 +80,9 @@ def skip_ball_ball_collision(rvw1, rvw2, s1, s2, R1, R2):
 
 
 @jit(nopython=True, cache=const.use_numba_cache)
-def get_u(rvw, R, phi, s):
+def get_u(
+    rvw: NDArray[np.float64], R: float, phi: float, s: int
+) -> NDArray[np.float64]:
     if s == const.rolling:
         return np.array([1, 0, 0], dtype=np.float64)
 
@@ -83,7 +94,19 @@ def get_u(rvw, R, phi, s):
 
 
 @jit(nopython=True, cache=const.use_numba_cache)
-def ball_ball_collision_coeffs(rvw1, rvw2, s1, s2, mu1, mu2, m1, m2, g1, g2, R):
+def ball_ball_collision_coeffs(
+    rvw1: NDArray[np.float64],
+    rvw2: NDArray[np.float64],
+    s1: int,
+    s2: int,
+    mu1: float,
+    mu2: float,
+    m1: float,
+    m2: float,
+    g1: float,
+    g2: float,
+    R: float,
+) -> Tuple[float, float, float, float, float]:
     """Get quartic coeffs required to determine the ball-ball collision time
 
     (just-in-time compiled)
@@ -139,7 +162,19 @@ def ball_ball_collision_coeffs(rvw1, rvw2, s1, s2, mu1, mu2, m1, m2, g1, g2, R):
     return a, b, c, d, e
 
 
-def ball_ball_collision_time(rvw1, rvw2, s1, s2, mu1, mu2, m1, m2, g1, g2, R):
+def ball_ball_collision_time(
+    rvw1: NDArray[np.float64],
+    rvw2: NDArray[np.float64],
+    s1: int,
+    s2: int,
+    mu1: float,
+    mu2: float,
+    m1: float,
+    m2: float,
+    g1: float,
+    g2: float,
+    R: float,
+) -> float:
     """Get the time until collision between 2 balls
 
     NOTE This is deprecated. Rather than solve the roots of a single polynomial
@@ -157,53 +192,20 @@ def ball_ball_collision_time(rvw1, rvw2, s1, s2, mu1, mu2, m1, m2, g1, g2, R):
 
 
 @jit(nopython=True, cache=const.use_numba_cache)
-def skip_ball_linear_cushion_collision(rvw, s, u_r, g, R, p1, p2, normal):
-    if s == const.spinning or s == const.pocketed or s == const.stationary:
-        # Ball isn't moving. No collision.
-        return True
-
-    if s == const.rolling:
-        # Since the ball is rolling, it is a straight line trajectory. The strategy here
-        # is to see whether the trajectory of the ball is going to intersect with either
-        # of the collisions defined by a linear cushion segment. Let r1 be the position
[…]
