### 2024-08-25 d2a37d8
Don't show cue stick during wait screen
