### 2024-08-31 #173 @ekiefl
# Update 1

Hi everyone,

Just wanted to give everyone an update on the slow (but at least somewhat consistent) progress.

## Implemented a wait screen when compiling/loading the numba function cache

Due to pooltool compiling numba functions (described in [this comment](https://github.com/pyOpenSci/software-submission/issues/173#issuecomment-2232444708)), there was sometimes a large delay when calculating the first shot. This gave the appearance that the shot was taking forever to calculate, when really, it was just one-time numba function compiling preamble. To separate this from the shot calculation itself, I made the following change: when the user first enters the interactive interface, a dummy shot is calculated in the background, during which time a loading screen appears explaining the wait time.

https://github.com/ekiefl/pooltool/pull/136

It looks like this:

## Event caching

@sebastianotronto had the idea of having an event cache so the same event isn't re-calculated at each step of the simulation. This was a very good idea and I'm happy to report some significant gains in speed. The details can be found in https://github.com/ekiefl/pooltool/pull/133, but here are speed improvements:

The results yield improvements in simulation time for random simulations, using ball number as a parameter for simulation complexity:

The improvement is more dramatic when the number of balls is dramatically increased:

I'm happy that caching _always_ seems worth it, regardless of ball number.

## Other

- Tests have been moved out of the package into `tests/` (https://github.com/ekiefl/pooltool/pull/135)
- Types were added to all the numba functions (https://github.com/ekiefl/pooltool/pull/137)
- Fixed code coverage badge (https://github.com/ekiefl/pooltool/commit/6b7c6889813657a14dbe4a03c7047e2dfdc96c40)

## Next

I've updated my above to-do checklist and will next be working on adding vignettes to the docs. Thank you for everyone's patience.
