+++ docs/examples/index.md
@@ -1,0 +1,11 @@
+# Examples
+
+## Contents
+
+```{eval-rst}
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   30_degree_rule
+```
