+++ docs/examples/30_degree_rule.pct.py
@@ -1,18 +1,120 @@
 # %% [markdown]
-# # 30-degree rule
+# # The 30 Degree Rule
 #
-# """This script validates the 30-degree rule proposed by Dr. Dave Billiards
+# The 30 degree rule states that the cue ball, when colliding with a ball over a wide range of cut angles, will be deflected roughly 30 degrees from it's initial course after the collision. In this sense, it is more of a *rule of thumb* used by pool players to improve their game, rather than a truism of pool physics.
 #
-# For more information about the 30-degree rule, see https://billiards.colostate.edu/faq/30-90-rules/30-degree-rule/
-# """
+# In this example, we will setup simulations that test the 30 degree rule and some of the physics equations defined by [Dr. Dave Billiards](https://drdavebilliards.com/).
+#
+# ## Assumptions
+#
+# We will use the default pooltool physics, which assume perfectly elastic and frictionless ball-ball collisions. Read more [here](https://ekiefl.github.io/2020/04/24/pooltool-theory/#section-ii-ball-ball-interactions).
+#
+# ## Definitions
+#
+# **The rule, stated in full:***
+#
+# > The 30° rule states that for a rolling-CB shot, over a wide range of cut angles, between a 1/4-ball hit (49 degree cut) and 3/4-ball hit (14 degree cut), the CB will deflect or carom off by very close to 30° (the “natural angle“) from its original direction after hitting the OB. If you want to be more precise, the angle is a little more (about 34°) closer to a 1/2-ball hit and a little less (about 27°) closer to a 1/4-ball or 3/4-ball hit.
+#
+# *(source: https://billiards.colostate.edu/faq/30-90-rules/30-degree-rule/)*
+#
+# **Ball-hit fraction and cut angle**
+#
+# - Ball-hit fraction, $f$, describes the fraction of overlap between the cue ball and object ball, projected in the direction of the aiming line.
+# - Cut angle, $\phi$, refers to the angle that the cue ball glances the object ball, where $0$ refers to a full ball hit (straight on), and $90$ refers to the lower bound of the thinnest hit possible.
+#
+# These two are visualized in this diagram, where $f = \text{ball overlap} / (2R)$:
+#
+# <img src="assets/30_degree_rule/diagram1.png" width="500px" />
+#
+# *(source: https://billiards.colostate.edu/technical_proofs/new/TP_A-23.pdf)*
+#
+# Establishing the relationship between these quantities is important, since the 30 degree rule makes reference to both cut angle *and* ball-hit fraction. One can calculate the ball-hit fraction from cut angle with the following equation:
+#
+# $$
+# f(\phi) = 1 - \sin{\phi}
+# $$
+
+# %% [markdown]
+# # Setting up the system
+#
+# To start, we'll need to create a billiards system. That means defining a table, a cue stick, and a collection of balls.
+#
+# We'll start with a table. Since we don't want collisions with cushions to interfere with our trajectory, let's make an unrealistically large $5\text{m} \times 5\text{m}$  table.
+
+# %%
+import pooltool as pt
+
+table_specs = pt.objects.BilliardTableSpecs(l=5, w=5)
+table = pt.Table.from_table_specs(table_specs)
+
+# %% [markdown]
+# Next, we'll create two balls.
+
+# %%
+cue_ball = pt.Ball.create("cue", xy=(2.5, 2.0))
+obj_ball = pt.Ball.create("obj", xy=(2.5, 3.0))
+
+# %% [markdown]
[…]
