+++ docs/examples/README.md
@@ -1,0 +1,7 @@
+# Summary
+
+All the examples are jupyter notebooks represented in percent format: https://jupytext.readthedocs.io/en/latest/formats-scripts.html#the-percent-format
+
+They can be opened as notebooks by running `jupyter lab`, right clicking the file, and selecting "Open as notebook"
+
+If you have a notebook, and would like to add it to the documentation, convert it to the percent format `jupytext --sync <yours.ipynb>`. It must be suffixed with `.pct.py`. Then open `index.md` and add it.
