### 2024-09-02 #173 @sebastianotronto
Great job @ekiefl ! I am glad that caching gives a performance improvement :)
