### 2024-09-02 69f9deb
Complete 30 deg rule
