+++ docs/examples/30_degree_rule.pct.py
@@ -1,4 +1,30 @@
-# %% [markdown]
+# ---
+# jupyter:
+#   jupytext:
+#     cell_metadata_filter: all
+#     notebook_metadata_filter: all
+#     text_representation:
+#       extension: .py
+#       format_name: percent
+#       format_version: '1.3'
+#       jupytext_version: 1.16.4
+#   kernelspec:
+#     display_name: Python 3 (ipykernel)
+#     language: python
+#     name: python3
+#   language_info:
+#     codemirror_mode:
+#       name: ipython
+#       version: 3
+#     file_extension: .py
+#     mimetype: text/x-python
+#     name: python
+#     nbconvert_exporter: python
+#     pygments_lexer: ipython3
+#     version: 3.12.4
+# ---
+
+# %% [markdown] editable=true slideshow={"slide_type": ""} tags=[]
 # # The 30 Degree Rule
 #
 # ## Intro
@@ -49,7 +75,7 @@
 #
 # We'll start with a table. Since we don't want collisions with cushions to interfere with our trajectory, let's make an unrealistically large $10\text{m} \times 10\text{m}$  [Table](../autoapi/pooltool/index.rst#pooltool.Table).
 
-# %%
+# %% trusted=true
 import pooltool as pt
 
 table_specs = pt.objects.BilliardTableSpecs(l=10, w=10)
@@ -59,20 +85,20 @@ table = pt.Table.from_table_specs(table_specs)
 # %% [markdown]
 # Next, we'll create two [Ball](../autoapi/pooltool/index.rst#pooltool.Ball) objects.
 
-# %%
+# %% trusted=true
 cue_ball = pt.Ball.create("cue", xy=(2.5, 1.5))
 obj_ball = pt.Ball.create("obj", xy=(2.5, 3.0))
 
 # %% [markdown]
 # Next, we'll need a [Cue](../autoapi/pooltool/index.rst#pooltool.Cue).
 
-# %%
+# %% trusted=true
 cue = pt.Cue(cue_ball_id="cue")
 
 # %% [markdown]
 # Finally, we'll need to wrap these objects up into a [System](../autoapi/pooltool/index.rst#pooltool.System). We'll call this our system *template*, with the intention of reusing it for many different shots.
 
-# %%
+# %% trusted=true
 system_template = pt.System(
     table=table,
     cue=cue,
@@ -86,7 +112,7 @@ system_template = pt.System(
 #
 # So in the function call below, `pt.aim.at_ball(system, "obj", cut=30)` returns the angle `phi` that the cue ball should be directed at such that a cut angle of 30 degrees with the object ball is achieved.
 
-# %%
+# %% trusted=true
 # Creates a deep copy of the template
 system = system_template.copy()
 
@@ -96,7 +122,7 @@ system.cue.set_state(V0=3, phi=phi, b=0.8)
 # %% [markdown]
 # Now, we [simulate](../autoapi/pooltool/index.rst#pooltool.simulate) the shot and then [continuize](../autoapi/pooltool/evolution/continuize/index.html#pooltool.evolution.continuize.continuize) it to store ball state data (like coordinates) in $10\text{ms}$ timestep intervals.
 
-# %%
+# %% trusted=true
 pt.simulate(system, inplace=True)
 pt.continuize(system, dt=0.01, inplace=True)
 
@@ -115,7 +141,7 @@ print(f"System simulated: {system.simulated}")
 #
 # Since that can't be embedded into the documentation, we'll instead plot the trajectory of the cue ball and object ball by accessing ther historical states.
 
-# %%
+# %% trusted=true
 cue_ball = system.balls["cue"]
 obj_ball = system.balls["obj"]
[…]
