### 2024-09-11 e5c4173
Throttle english only in the interactive interface
