### 2024-09-12 92317b2
Update default physics to include squirt
