+++ docs/examples/straight_shot.pct.py
@@ -1,0 +1,313 @@
+# ---
+# jupyter:
+#   jupytext:
+#     notebook_metadata_filter: all
+#     text_representation:
+#       extension: .py
+#       format_name: percent
+#       format_version: '1.3'
+#       jupytext_version: 1.16.4
+#   kernelspec:
+#     display_name: Python 3 (ipykernel)
+#     language: python
+#     name: python3
+#   language_info:
+#     codemirror_mode:
+#       name: ipython
+#       version: 3
+#     file_extension: .py
+#     mimetype: text/x-python
+#     name: python
+#     nbconvert_exporter: python
+#     pygments_lexer: ipython3
+#     version: 3.12.4
+# ---
+
+# %% [markdown] editable=true slideshow={"slide_type": ""}
+# # What straight shot is most difficult?
+#
+# Consider a straight shot into a pocket where the cue ball (CB) is a total distance $D$ from the pocket. Which object ball (OB) position (distance $d$ from the CB) results in the most difficult pot?
+#
+# This question is posed by Dr. Dave Billiards in [this proof](https://billiards.colostate.edu/technical_proofs/TP_3-4.pdf) and based on the assumptions made, the answer is
+#
+# $$
+# d = D/2
+# $$
+#
+# In other words, the shot is made most difficult when the OB is placed equidistantly from both the pocket and the CB.
+#
+# In this example, we will directly simulate straight shots and determine under what circumstances this theory holds up.
+
+# %% [markdown]
+# ## Setting up a system
+#
+# Let's set up a system where the OB is shot into a corner pocket _straight on_ (at an angle of 45 degrees from both the long cushion and short cushion).
+#
+# We'll begin by creating a big square [Table](../autoapi/pooltool/index.rst#pooltool.Table) that gives us plenty of space. There are many different parameters that can be passed to [PocketTableSpecs](../autoapi/pooltool/objects/index.rst#pooltool.objects.PocketTableSpecs) that influence the shape of the pockets (you can get a visual explanation of each parameter [here](../resources/table_specs.md)), but here we will keep the default values, which are modelled after my old 7-foot Showood table with pretty wide 5 inch pockets.
+
+# %%
+import pooltool as pt
+
+table_specs = pt.objects.PocketTableSpecs(l=10, w=10)
+table = pt.Table.from_table_specs(table_specs)
+
+# %% [markdown]
+# As a reminder, $D$ represents the distance of the CB to the pocket, but let's be a little more precise with what that means. From the CB, we will measure from the ball's center, and from the pocket we will measure from pocket's edge that's closest to the CB. So when $D$ is 0, the CB is teetering on the edge of the pocket.
+#
+# With that in mind, let's pick the bottom-left corner [Pocket](../autoapi/pooltool/objects/index.rst#pooltool.objects.Pocket) as our target and place a CB a distance $D$ away.
+
+# %%
+import numpy as np
+
+# lb stands for left-bottom
+pocket = table.pockets["lb"]
+
+# The pocket's center is defined with 3D coordinates, but we'll just use X and Y
+center = pocket.center[:2]
+
+pocket_point = center + pocket.radius * np.sqrt(2)
+
+print(f"The pocket point is {pocket_point}")
+
+# %% [markdown]
+# Now let's create a cue ball a distance $D$ away.
+
+# %%
+D = 1.0
+
+def create_cue_ball(pocket_point, D):
+    cue_point = pocket_point + D * np.sqrt(2)
+    return pt.Ball.create("CB", xy=cue_point)
+
+cue_ball = create_cue_ball(pocket_point, D)
+
+
[…]
