### 2024-09-14 5819ae7
Finish straight shot example
- Add docs to Makefile
