+++ docs/examples/index.md
@@ -3,5 +3,6 @@
 ```{eval-rst}
 .. nbgallery::
    30_degree_rule
+   straight_shot
 ```
 
