+++ paper/paper.bib
@@ -934,3 +934,43 @@
   howpublished = {\url{https://github.com/ekiefl/FastFiz}},
   note = {Accessed: 2024-04-03}
 }
+
+@software{siu_kwan_lam_2024_11642058,
+  author       = {Siu Kwan Lam and
+                  stuartarchibald and
+                  Antoine Pitrou and
+                  Mark Florisson and
+                  Graham Markall and
+                  Stan Seibert and
+                  Emergency Self-Construct and
+                  Todd A. Anderson and
+                  Guilherme Leobas and
+                  rjenc29 and
+                  Michael Collison and
+                  luk-f-a and
+                  Kaustubh and
+                  Jay Bourque and
+                  Aaron Meurer and
+                  Travis E. Oliphant and
+                  Nick Riasanovsky and
+                  Michael Wang and
+                  densmirn and
+                  KrisMinchev and
+                  Andre Masella and
+                  Ethan Pronovost and
+                  njwhite and
+                  Eric Wieser and
+                  Ehsan Totoni and
+                  Stefan Seefeld and
+                  Hernan Grecco and
+                  Pearu Peterson and
+                  Isaac Virshup and
+                  MattyG},
+  title        = {numba/numba: 0.60.0},
+  month        = jun,
+  year         = 2024,
+  publisher    = {Zenodo},
+  version      = {0.60.0},
+  doi          = {10.5281/zenodo.11642058},
+  url          = {https://doi.org/10.5281/zenodo.11642058}
+}
