### 2024-09-14 265759b
Fix links with new MyST cross-referencing
