+++ docs/getting_started/script.md
@@ -36,13 +36,13 @@ For those interested in some exposition, below is a line-by-line explanation.
 
 ## Explanation
 
-First thing first, the pootool package is imported which gives you access to the [top-level API](#pooltool):
+First thing first, the pootool package is imported which gives you access to the {py:mod}`top-level API <pooltool>`:
 
 ```python
 import pooltool as pt
 ```
 
-Then a table, a cue stick, and a collection of balls are created and wrapped up into a [](#pooltool.system.datatypes.System):
+Then a table, a cue stick, and a collection of balls are created and wrapped up into a {py:class}`System <pooltool.System>`:
 
 ```python
 # We need a table, some balls, and a cue stick
@@ -54,7 +54,7 @@ cue = pt.Cue(cue_ball_id="cue")
 shot = pt.System(table=table, balls=balls, cue=cue)
 ```
 
-The cue stick parameters are then set with [](#pooltool.objects.cue.datatypes.Cue.set_state). A large impact speed of `V0=8` (m/s) is chosen, and an aiming utility function is used to aim the cue ball directly at the one-ball.
+The cue stick parameters are then set with {py:meth}`pooltool.Cue.set_state`. A large impact speed of `V0=8` (m/s) is chosen, and an aiming utility function is used to aim the cue ball directly at the one-ball.
 
 ```python
 # Aim at the head ball with a strong impact
@@ -66,12 +66,12 @@ shot.cue.set_state(V0=8, phi=pt.aim.at_ball(shot, "1"))
 A description of the cue's parameters can be found in the API docs:
 
 ```{eval-rst}
-.. autoclass:: pooltool.objects.cue.datatypes.Cue
+.. autoclass:: pooltool.Cue
     :noindex:
 ```
 :::
 
-The cue parameters have been set, but the system still hasn't been simulated. This is done with a call to [](#pooltool.evolution.event_based.simulate.simulate):
+The cue parameters have been set, but the system still hasn't been simulated. This is done with a call to {py:func}`pooltool.simulate`.
 
 ```python
 # Evolve the shot.
@@ -80,7 +80,7 @@ pt.simulate(shot, inplace=True)
 
 The system has now been evolved from its initial to its final state.
 
-To visualize the shot, open the GUI with [](#pooltool.interact.ShotViewer):
+To visualize the shot, open the GUI with {py:class}`pooltool.ShotViewer`:
 
 ```python
 # Open up the shot in the GUI
@@ -90,5 +90,4 @@ interface.show(shot)
 
 ## Next
 
-Obviously this script is just the beginning. Pooltool offers much more than this, which means you have much more to learn. From here, I suggest you check out the [Examples](examples) page and dive into whatever topic interests you.
-
+Obviously this script is just the beginning. Pooltool offers much more than this, which means you have much more to learn. From here, I suggest you check out the [Examples](../examples/index.md) page and dive into whatever topic interests you.
