### 2024-09-14 #173 @ekiefl
# Update 2

We're getting *really* close.

## Established vignette doc infrastructure

I think jupyter notebooks provide an great hands-on approach to learning, so I really wanted the vignettes to be rendered as notebooks. This was made possible by [nbsphinx](https://nbsphinx.readthedocs.io/en/0.9.5/), and the infrastructure for adding these to the docs was done in https://github.com/ekiefl/pooltool/pull/138.

Some details about this:

1. No .ipynb files are git-tracked. Instead, they are converted to a [python percent](https://jupytext.readthedocs.io/en/latest/formats-scripts.html#the-percent-format) format using [jupytext](https://jupytext.readthedocs.io/en/latest). This makes line diffs clearer, reduces file size, and enables python linting on notebooks.
2. The notebooks are executed when the docs are built, and then rendered as HTML.

## Added two vignettes

A new section of the documentation now exists, called [Examples](https://pooltool.readthedocs.io/en/latest/examples/index.html). It displays a gallery of all the notebook examples/vignettes. Currently there are two.

The first is an analysis of the so-called "30 Degree Rule" and was added in https://github.com/ekiefl/pooltool/pull/138.

The second is an analysis of shot difficulty for straight in shots and was added in https://github.com/ekiefl/pooltool/pull/140

## Updated the paper

I've updated the `paper.md`:

1. There is now a figure displaying screenshots of the interface
2. I added a section called "Implementation" that gives implementation details about the core features of pooltool
3. I added a "Usage" section that itself contains no code, but references the pooltool documentation. I did not want to include any pooltool code in the paper because the API is subject to change and the paper content is permanent. Although code usage is seen in other JOSS papers, it goes against the [JOSS guidelines](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain), which state "_software documentation such as API (Application Programming Interface) functionality should not be in the paper and instead should be outlined in the software documentation_"

These changes are found in https://github.com/ekiefl/pooltool/pull/141

Here is a rendered version of the paper: [paper.zip](https://github.com/user-attachments/files/17004036/paper.zip)

## What next?

This morning my CI tests are crashing on the Linux builds, so I'll resolve that in due time before merging https://github.com/ekiefl/pooltool/pull/141.

Other than that, my TODO list is, for the time being, **complete**!

@sebastianotronto and @eliotwrobson, could you please double check that your checklists are up to date? If there are any remaining boxes to be ticked, could you please make a suggestion for what you would like to see done?
