+++ .github/workflows/test.yml
@@ -11,7 +11,7 @@ jobs:
       matrix:
         python-version: ["3.9", "3.12"]
         poetry-version: ["1.8.3"]
-        os: [ubuntu-latest, macos-latest, windows-latest]
+        os: [ubuntu-20.04, macos-latest, windows-latest]
     runs-on: ${{ matrix.os }}
     steps:
       - uses: actions/checkout@v4
@@ -38,7 +38,7 @@ jobs:
 
       # Calculate the test coverage and upload results
       - name: Get and upload test coverage
-        if: matrix.python-version == '3.12' && matrix.os == 'ubuntu-latest'
+        if: matrix.python-version == '3.12' && matrix.os == 'ubuntu-20.04'
         run: |
           # Replace `linux` below with the appropriate OS (Options are `alpine`, `linux`, `macos`, `windows`)
           poetry run pytest --cov pooltool --cov-report=xml
