+++ .github/workflows/test.yml
@@ -35,7 +35,7 @@ jobs:
         run: poetry install
         env:
           # Don't give up too soon
-          PIP_DEFAULT_TIMEOUT: '120'
+          POETRY_REQUESTS_TIMEOUT: '120'
       - name: Run tests
         run: poetry run pytest
 
