### 2024-09-15 f9a70cc
Handle ShotViewer instantiation internally
