### 2024-09-15 640adbd
Big lag if window not removed from graphicsEngine
