### 2024-09-15 709e07a
Minimization
- Not great
- Scene doesn't stop I dont think
