### 2024-09-15 4ced109
Window now closes when `esc` is pressed
