### 2024-09-15 eb021cd
Window now closes when `esc` is pressed
