### 2024-09-15 9bb92fc
Fix game mode
