### 2024-09-16 #173 @cmarmo
Congratulations @ekiefl !

🎉 Pooltool has been approved by pyOpenSci! Thank you for submitting Pooltool and many thanks to @sebastianotronto and @eliotwrobson  for reviewing this package! 😸

- [ ] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

<IF JOSS SUBMISSION>

- [ ] *Editor Task* Once the JOSS issue is opened for the package, we strongly suggest that you subscribe to issue updates. This will allow you to continue to update the issue labels on this review as it goes through the JOSS process.

<IF JOSS SUBMISSION/>

---
