### 2024-09-16 #173 @cmarmo
Thank you to all of you for all your work!
@ekiefl please let me know if you are willing to connect with the pyOpenSci community on our slack, I can send you an invite.
Also, if you are planning to describe your review experience in a blog post let us know so we can cross post on our socials.
