### 2024-09-21 #173 @ekiefl
@cmarmo, @eliotwrobson, @sebastianotronto, thank you *so* much for volunteering your time. Thanks to all of your efforts, pooltool has never been in better shape.

This is fantastic news!

> @ekiefl please let me know if you are willing to connect with the pyOpenSci community on our slack, I can send you an invite.

Please do.

> Also, if you are planning to describe your review experience in a blog post let us know so we can cross post on our socials.

I wasn't planning on it, but I'll let you know.

I'm going to work through the post-acceptance tasks now.
