### 2024-09-22 fb3d1cb
Fix test PyPI repo declaration
