### 2024-09-22 1f8dbd4
Use URL for logo.png
Otherwise the PyPI README will not render the image.
