### 2024-09-22 b702b58
Add pyOpenSci badge
