### 2024-09-22 b2158d0
Update README.md
Fix pyOpenSci link
