+++ paper/paper.bib
@@ -1,133 +1,3 @@
-@ARTICLE{Pinzon2017-jv,
-  title     = "Billiard game parameters calculation using a depth camera for
-               augmented reality applications",
-  author    = "Pinzon, Diego Mauricio Rivera and Gonzalez, Edilberto Carlos
-               Vivas and Gomez, Edwar Jacinto",
-  journal   = "Contemp. Eng. Sci.",
-  publisher = "Hikari, Ltd.",
-  volume    =  10,
-  pages     = "1171--1180",
-  year      =  2017
-}
-
-@INPROCEEDINGS{Fu2022-vl,
-  title     = "Visual aided training of billiards based on depth estimation",
-  booktitle = "2022 {IEEE} Conference on Telecommunications, Optics and
-               Computer Science ({TOCS})",
-  author    = "Fu, Jiyuan and Wang, Weining and Chang, Lechang",
-  abstract  = "Billiards is a popular recreational sport, but because of the
-               strong technical movement, beginners can not choose the relevant
-               shot path. Given this situation, it is very promising and
-               meaningful to use computer vision to assist billiards. However,
-               due to the limited height and size of billiard ball, mobile
-               phone photography can only get a side view, unable to accurately
-               convey the cue ball's relative position and other balls.
-               Therefore, the white ball hitting path planning can not achieve
-               the ideal hitting effect in reality to assist users. This paper
-               proposes a method of calculating the relative position of the
-               cue ball and other balls by using monocular depth estimation and
-               restoring them to the top view. Traditional monocular depth
-               estimation requires the artificial assumption that the
-               relationship between RBG image and depth satisfies a certain
-               parameter model, which results in limited prediction accuracy
-               and consumes computational resources. In this paper, the
-               monocular prediction method based on deep learning is adopted,
-               and the combination of PVCNN and global feature extraction
-               method is proposed, which can improve the prediction accuracy
-               and reduce memory consumption. The experimental results show
-               that this method can be used to estimate the depth of
-               information of billiards, which lays a foundation for the
-               visual-assisted training of billiards based on mobile phones.",
-  publisher = "IEEE",
-  pages     = "397--403",
-  month     =  dec,
-  year      =  2022
-}
-
-@INPROCEEDINGS{Fu2022-ww,
-  title     = "Visual aided training of billiards based on depth estimation",
-  booktitle = "2022 {IEEE} Conference on Telecommunications, Optics and
-               Computer Science ({TOCS})",
-  author    = "Fu, Jiyuan and Wang, Weining and Chang, Lechang",
-  abstract  = "Billiards is a popular recreational sport, but because of the
-               strong technical movement, beginners can not choose the relevant
-               shot path. Given this situation, it is very promising and
-               meaningful to use computer vision to assist billiards. However,
-               due to the limited height and size of billiard ball, mobile
-               phone photography can only get a side view, unable to accurately
-               convey the cue ball's relative position and other balls.
-               Therefore, the white ball hitting path planning can not achieve
-               the ideal hitting effect in reality to assist users. This paper
-               proposes a method of calculating the relative position of the
-               cue ball and other balls by using monocular depth estimation and
-               restoring them to the top view. Traditional monocular depth
-               estimation requires the artificial assumption that the
[…]
