+++ paper/paper.bib
@@ -54,7 +54,7 @@
   pages     = "924--932",
   month     =  dec,
   year      =  2019,
-  doi       =  {"10.1109/SSCI44817.2019.9003085"}
+  doi       =  "10.1109/SSCI44817.2019.9003085"
 }
 
 @INPROCEEDINGS{Nierhoff2012-st,
