+++ paper/paper.bib
@@ -54,7 +54,7 @@
   pages     = "924--932",
   month     =  dec,
   year      =  2019,
-  doi       =  "10.1109/SSCI44817.2019.9003085"
+  doi       =  {10.1109/SSCI44817.2019.9003085}
 }
 
 @INPROCEEDINGS{Nierhoff2012-st,
@@ -68,7 +68,7 @@
   conference      = "2012 IEEE Workshop on Advanced Robotics and its Social
                      Impacts (ARSO)",
   location        = "Munich",
-  doi             = {"10.1109/ARSO.2012.6213402"}
+  doi             = {10.1109/ARSO.2012.6213402}
 }
 
 @ARTICLE{Alian2004-zs,
@@ -117,7 +117,7 @@
   pages     = "184--195",
   month     =  feb,
   year      =  2016,
-  doi       = {"10.1109/TMECH.2015.2461547"}
+  doi       = {10.1109/TMECH.2015.2461547}
 }
 
 @ARTICLE{Greenspan2008-wg,
@@ -139,7 +139,7 @@
   pages     = "46--53",
   month     =  jan,
   year      =  2008,
-  doi       = {"10.1109/MC.2008.33"}
+  doi       = {10.1109/MC.2008.33}
 }
 
 @ARTICLE{Sousa2016-vi,
@@ -164,7 +164,7 @@
   number    =  2,
   month     =  apr,
   year      =  2016,
-  doi       = {"10.1007/s41095-016-0047-3"}
+  doi       = {10.1007/s41095-016-0047-3}
 }
 
 @ARTICLE{Smith2007-jq,
@@ -193,7 +193,7 @@
   pages    = "1069--1091",
   month    =  nov,
   year     =  2007,
-  doi      = {"10.1016/j.artint.2007.04.011"},
+  doi      = {10.1016/j.artint.2007.04.011},
   keywords = "Game-tree search; Computer games; Uncertainty; Monte-Carlo
               methods; Billiards"
 }
@@ -241,7 +241,7 @@
   pages    = "21543--21575",
   month    =  oct,
   year     =  2023,
-  doi      = {"10.1007/s10489-023-04542-3"}
+  doi      = {10.1007/s10489-023-04542-3}
 }
 
 @ARTICLE{Fragkiadaki2015-oh,
@@ -272,7 +272,7 @@
   archivePrefix = "arXiv",
   primaryClass  = "cs.CV",
   eprint        = "1511.07404",
-  doi           = {"10.48550/arXiv.1511.07404"}
+  doi           = {10.48550/arXiv.1511.07404}
 }
 
 @ARTICLE{Silva2018-cm,
@@ -292,7 +292,7 @@
   publisher = "Revista de Ci{\^e}ncias da Computa{\c c}{\~a}o",
   year      =  2018,
   language  = "en",
-  doi       = {"10.34627/RCC.V12IESPECIAL.13"}
+  doi       = {10.34627/RCC.V12IESPECIAL.13}
 }
 
 @INPROCEEDINGS{Chen2019-dk,
@@ -316,7 +316,7 @@
   pages     = "1--4",
   month     =  aug,
   year      =  2019,
-  doi       = {"10.1109/CIG.2019.8848113"}
+  doi       = {10.1109/CIG.2019.8848113}
 }
 
 @ARTICLE{Archibald2016-sd,
@@ -345,7 +345,7 @@
   pages     = "190--202",
   month     =  jun,
   year      =  2016,
-  doi       = {"10.1109/TCIAIG.2016.2549748"}
+  doi       = {10.1109/TCIAIG.2016.2549748}
 }
 
 % The entry below contains non-ASCII chars that could not be converted
[…]
