### 2024-09-28 86579a7
All lowercase DOIs
