### 2024-09-28 6be6a11
Remove all italicizing braces
