+++ paper/paper.bib
@@ -30,10 +30,10 @@
 }
 
 @INPROCEEDINGS{Tung2019-zu,
-  title     = "Toward Human-like Billiard {AI} Bot Based on Backward Induction
+  title     = "Toward Human-like Billiard AI Bot Based on Backward Induction
                and Machine Learning",
-  booktitle = "2019 {IEEE} Symposium Series on Computational Intelligence
-               ({SSCI})",
+  booktitle = "2019 IEEE Symposium Series on Computational Intelligence
+               (SSCI)",
   author    = "Tung, Kuei Gu and Wang, Sheng Wen and Tai, Wen Kai and Way, Der
                Lor and Chang, Chin Chen",
   abstract  = "A human-like billiard AI bot approach is proposed in this paper.
@@ -59,8 +59,8 @@
 
 @INPROCEEDINGS{Nierhoff2012-st,
   title           = "Strategic play for a pool-playing robot",
-  booktitle       = "2012 {IEEE} Workshop on Advanced Robotics and its Social
-                     Impacts ({ARSO})",
+  booktitle       = "2012 IEEE Workshop on Advanced Robotics and its Social
+                     Impacts (ARSO)",
   author          = "Nierhoff, T and Heunisch, K and Hirche, S",
   publisher       = "IEEE",
   month           =  may,
@@ -71,17 +71,20 @@
   doi             = {10.1109/arso.2012.6213402}
 }
 
-@ARTICLE{Alian2004-zs,
-  title   = "Roboshark: A gantry pool player robot",
-  author  = "Alian, M E and Shouraki, S B and Shalmani, M and Karimian, P and
-             {others}",
-  journal = "Thirty-Fifth International",
-  year    =  2004
+
+% There appears to be no one single source for this piece of work. This source
+% used below is from this Google Scholar entry:
+% https://scholar.google.com/scholar?cluster=9435098466170805552&hl=en&oi=scholarr
+@INPROCEEDINGS{Alian2004-zs,
+  title={Roboshark: A gantry pool player robot},
+  author={Alian, Mohammad Ebne and Shouraki, Saeed Bagheri and Shalmani, M and Karimian, Pooya and Sabzmeydani, Payam},
+  booktitle={Thirty-Fifth International Symposium on Robotics (ISR 2004), Paris},
+  year={2004}
 }
 
 @ARTICLE{Mathavan2016-ck,
   title     = "Ball Positioning in Robotic Billiards: A Nonprehensile
-               {Manipulation-Based} Solution",
+               Manipulation-Based Solution",
   author    = "Mathavan, Senthan and Jackson, Michael R and Parkin, Robert M",
   abstract  = "The last two decades have seen a number of developments in
                creating robots to play billiards. The designed robotic systems
@@ -121,7 +124,7 @@
 }
 
 @ARTICLE{Greenspan2008-wg,
-  title     = "Toward a Competitive {Pool-Playing} Robot",
+  title     = "Toward a Competitive Pool-Playing Robot",
   author    = "Greenspan, Michael and Lam, Joseph and Godard, Marc and Zaidi,
                Imran and Jordan, Sam and Leckie, Will and Anderson, Ken and
                Dupuis, Donna",
@@ -168,7 +171,7 @@
 }
 
 @ARTICLE{Smith2007-jq,
-  title    = "{PickPocket}: A computer billiards shark",
+  title    = "PickPocket: A computer billiards shark",
   author   = "Smith, Michael",
   abstract = "Billiards is a game of both strategy and physical skill. To
               succeed, a player must be able to select strong shots, and then
@@ -199,7 +202,7 @@
 }
 
 @ARTICLE{Rodriguez-Lozano2023-hq,
-  title    = "{3D} reconstruction system and multiobject local tracking
[…]
