### 2024-09-28 b147dd7
Properly cite masters and phd theses
