+++ paper/paper.bib
@@ -1,32 +1,16 @@
-@BOOK{Bhagat2018-bx,
-  title     = "Automatic Snooker-playing Robot with Speech Recognition Using
-               Deep Learning",
-  author    = "Bhagat, Kunj H",
-  abstract  = "Abstract: Research on natural language processing, such as for
-               image and speech recognition, is rapidly changing focus from
-               statistical methods to neural networks. In this study, we
-               introduce speech recognition capabilities along with computer
-               vision to allow a robot to play snooker completely by itself.
-               The color of the ball to be pocketed is provided as an audio
-               input using an audio device such as a microphone. The system is
-               able to recognize the color from the input using a trained deep
-               learning network. The system then commands the camera to locate
-               the ball of the identified color on a snooker table by using
-               computer vision. To pocket the target ball, the system then
-               predicts the best shot using an algorithm. This activity can be
-               executed accurately based on the efficiency of the trained deep
-               learning model.",
-  publisher = "California State University, Long Beach",
-  year      =  2018,
-  language  = "en"
+@MASTERSTHESIS{Bhagat2018-bx,
+  title  = "Automatic Snooker-playing Robot with Speech Recognition Using
+            Deep Learning",
+  author = "Bhagat, Kunj H",
+  school = "California State University, Long Beach",
+  year   =  2018
 }
 
-@BOOK{Sang1994-jv,
-  title     = "Automating Skills Using a Robot Snooker Player",
-  author    = "Sang, William Cheung Shu",
-  publisher = "University of Bristol",
-  year      =  1994,
-  language  = "en"
+@PHDTHESIS{Sang1994-jv,
+  title  = "Automating Skills Using a Robot Snooker Player",
+  author = "Sang, William Cheung Shu",
+  school = "University of Bristol",
+  year   =  1994
 }
 
 @INPROCEEDINGS{Tung2019-zu,
