### 2024-09-29 #173 @ekiefl
Hi @cmarmo, I wanted to update you that pooltool has been accepted into JOSS 🥳

https://github.com/openjournals/joss-reviews/issues/7301
