### 2024-09-29 #173 @cmarmo
That's wonderful @ekiefl ! I'm going to announce it on our [discourse](https://pyopensci.discourse.group/)!

@eliotwrobson , @sebastianotronto may I ask you to fill our [post-review survey](https://forms.gle/BLGVCfUdiJHS5YJY7)? This is a very helpful feedback for us in order to improve our review process... thank you so much for your understanding!
